requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiofile==3.9.0
    # via py-key-value-aio
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   mcp
    #   py-key-value-aio
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==26.1.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.9
    # via fastmcp
backports-tarfile==1.2.0
    # via jaraco-context
beartype==0.22.9
    # via py-key-value-aio
cachetools==7.0.5
    # via py-key-value-aio
caio==0.9.25
    # via aiofile
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
contourpy==1.3.2
    # via matplotlib
cryptography==46.0.6
    # via
    #   authlib
    #   pyjwt
cycler==0.12.1
    # via matplotlib
cyclopts==4.10.1
    # via fastmcp
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.4
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via
    #   anyio
    #   fastmcp
    #   pytest
fastmcp==3.2.0
    # via -r requirements.in
filelock==3.25.2
    # via huggingface-hub
fonttools==4.62.1
    # via matplotlib
fsspec==2026.3.0
    # via huggingface-hub
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   huggingface-hub
    #   langchain-mistralai
    #   langgraph-sdk
    #   langsmith
    #   mcp
httpx-sse==0.4.3
    # via
    #   langchain-mistralai
    #   mcp
huggingface-hub==1.10.0
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via
    #   keyring
    #   opentelemetry-api
iniconfig==2.3.0
    # via pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.1.2
    # via keyring
jaraco-functools==4.4.0
    # via keyring
joblib==1.5.3
    # via
    #   -r requirements.in
    #   scikit-learn
joblib-stubs==1.5.2.0.20250831
    # via -r requirements.in
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
jsonref==1.1.0
    # via fastmcp
jsonschema==4.26.0
    # via mcp
jsonschema-path==0.4.5
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.7.0
    # via py-key-value-aio
kiwisolver==1.5.0
    # via matplotlib
langchain==1.2.15
    # via -r requirements.in
langchain-core==1.2.26
    # via
    #   langchain
    #   langchain-mcp-adapters
    #   langchain-mistralai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-mcp-adapters==0.2.2
    # via -r requirements.in
langchain-mistralai==1.1.2
    # via -r requirements.in
langgraph==1.1.6
    # via langchain
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.9
    # via langgraph
langgraph-sdk==0.3.12
    # via langgraph
langsmith==0.7.26
    # via langchain-core
markdown-it-py==4.0.0
    # via rich
matplotlib==3.10.8
    # via -r requirements.in
mcp==1.27.0
    # via
    #   fastmcp
    #   langchain-mcp-adapters
mdurl==0.1.2
    # via markdown-it-py
more-itertools==11.0.1
    # via
    #   jaraco-classes
    #   jaraco-functools
numpy==2.2.6
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   pandas-stubs
    #   scikit-learn
    #   scipy
openapi-pydantic==0.5.1
    # via fastmcp
opentelemetry-api==1.40.0
    # via fastmcp
orjson==3.11.8
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   fastmcp
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   matplotlib
    #   pytest
pandas==2.3.3
    # via -r requirements.in
pandas-stubs==2.3.3.260113
    # via -r requirements.in
pathable==0.5.0
    # via jsonschema-path
pillow==12.2.0
    # via matplotlib
platformdirs==4.9.4
    # via fastmcp
pluggy==1.6.0
    # via pytest
py-key-value-aio[filetree,keyring,memory]==0.4.4
    # via fastmcp
pycparser==3.0
    # via cffi
pydantic[email]==2.12.5
    # via
    #   fastmcp
    #   langchain
    #   langchain-core
    #   langchain-mistralai
    #   langgraph
    #   langsmith
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pygments==2.20.0
    # via
    #   pytest
    #   rich
pyjwt[crypto]==2.12.1
    # via mcp
pyparsing==3.3.2
    # via matplotlib
pyperclip==1.11.0
    # via fastmcp
pytest==9.0.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.24
    # via mcp
pytz==2026.1.post1
    # via pandas
pywin32==311
    # via mcp
pywin32-ctypes==0.2.3
    # via keyring
pyyaml==6.0.3
    # via
    #   fastmcp
    #   huggingface-hub
    #   jsonschema-path
    #   langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==14.3.3
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
    #   typer
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.7.2
    # via -r requirements.in
scikit-learn-stubs==0.0.3
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sse-starlette==3.3.4
    # via mcp
starlette==1.0.0
    # via
    #   mcp
    #   sse-starlette
tenacity==9.1.4
    # via langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via langchain-mistralai
tomli==2.4.1
    # via
    #   cyclopts
    #   pytest
tqdm==4.67.3
    # via huggingface-hub
typer==0.24.1
    # via huggingface-hub
types-pytz==2026.1.1.20260402
    # via pandas-stubs
typing-extensions==4.15.0
    # via
    #   anyio
    #   cryptography
    #   cyclopts
    #   exceptiongroup
    #   huggingface-hub
    #   joblib-stubs
    #   langchain-core
    #   langchain-mcp-adapters
    #   mcp
    #   opentelemetry-api
    #   py-key-value-aio
    #   pydantic
    #   pydantic-core
    #   pyjwt
    #   referencing
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via pandas
uncalled-for==0.3.0
    # via fastmcp
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.44.0
    # via
    #   -r requirements.in
    #   fastmcp
    #   mcp
watchfiles==1.1.1
    # via fastmcp
websockets==16.0
    # via fastmcp
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
